refactor(relay): テストデータビルダーをfixturesモジュールに分離
オブジェクトやsubscriptionのビルダーは振る舞いの代役（mock）ではなく
本物の型にテスト値を詰めるfixtureなので、mocksから独立させる。
検証ヘルパがupstream側mockに依存するねじれも解消。
