+++ relay/src/modules/relay/tests/data_plane.rs
@@ -12,11 +12,9 @@
 use bytes::Bytes;
 
 use super::{
+    fixtures::ordered_payload,
     harness::{DataPlaneHarness, OBJECT_COUNT, assert_full_ordered_delivery, collect_until_closed},
-    mocks::{
-        downstream_client::{header_group_on_stream, payloads_on_stream, stream_closed},
-        upstream_client::ordered_payload,
-    },
+    mocks::downstream_client::{header_group_on_stream, payloads_on_stream, stream_closed},
 };
 
 /// Baseline: egress is running (largest resolved as None before any object
