+++ relay/src/modules/relay/tests/mocks/downstream_client.rs
@@ -128,24 +128,6 @@ impl Publisher for MockPublisher {
     }
 }
 
-/// A LargestObject/Ascending downstream subscription, matching what the
-/// cascading-relay E2E subscriber sends.
-pub(crate) fn make_downstream_subscription() -> DownstreamSubscription {
-    DownstreamSubscription::from(moqt::Subscription::SubscriberInitiated(
-        moqt::SubscriberInitiatedSubscription {
-            request_id: 0,
-            track_namespace: "ns".to_string(),
-            track_name: "track".to_string(),
-            track_alias: 0,
-            expires: 0,
-            group_order: moqt::GroupOrder::Ascending,
-            content_exists: moqt::ContentExists::False,
-            filter_type: moqt::FilterType::LargestObject,
-            delivery_timeout: None,
-        },
-    ))
-}
-
 // ---------------------------------------------------------------------------
 // Queries over the recorded events
 // ---------------------------------------------------------------------------
